Extract modules `kompile` and `prove` (#51)
* Move `_init_and_run_proof` into `run_kcfg_group`

* Move `run_cfg_group` out of `foundry_prove`

* Rename functions

* Extract module `kontrol.prove`

* Rename function

* Extract module `kontrol.kompile`

* Move import to top

* Set Version: 0.1.9

---------
